# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ORN-AI Talent Infrastructure Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register a new user account (candidate, recruiter or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates with filters (recruiter/admin view)
      parameters:
        - name: country
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
        - name: experienceMin
          in: query
          schema: { type: integer }
        - name: experienceMax
          in: query
          schema: { type: integer }
        - name: englishLevel
          in: query
          schema: { type: string }
        - name: minReadiness
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"

    post:
      operationId: registerCandidate
      tags: [candidates]
      summary: Register a new candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateRegistration"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}/cv:
    post:
      operationId: uploadCv
      tags: [candidates]
      summary: Attach CV metadata to a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CvUpload"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
...
components:
  schemas:
...
    EnglishLevel:
      type: string
      enum: [A1, A2, B1, B2, C1, C2]
    VisaStatus:
      type: string
      enum: [eu_citizen, work_permit, blue_card, requires_sponsorship, student_visa]
    CandidateRegistration:
      type: object
      properties:
        fullName: { type: string, minLength: 2 }
        email: { type: string, format: email }
        phone: { type: string }
        country: { type: string }
        targetRole: { type: string }
        yearsExperience: { type: integer, minimum: 0, maximum: 50 }
        visaStatus: { $ref: "#/components/schemas/VisaStatus" }
        englishLevel: { $ref: "#/components/schemas/EnglishLevel" }
        euWorkEligible: { type: boolean }
        linkedinUrl: { type: string }
        skills:
          type: array
          items: { type: string }
      required:
        - fullName
        - email
        - phone
        - country
        - targetRole
        - yearsExperience
        - visaStatus
        - englishLevel
        - euWorkEligible
        - linkedinUrl
    CvUpload:
      type: object
      properties:
        fileName: { type: string }
        fileSize: { type: integer }
        contentSummary: { type: string }
      required: [fileName, fileSize]
    EvaluationScores:
      type: object
      properties:
        cvQuality: { type: integer, minimum: 0, maximum: 100 }
        technicalSkillMatch: { type: integer, minimum: 0, maximum: 100 }
        technicalRelevance: { type: integer, minimum: 0, maximum: 100 }
        englishReadiness: { type: integer, minimum: 0, maximum: 100 }
        europeJobReadiness: { type: integer, minimum: 0, maximum: 100 }
        marketReadiness: { type: integer, minimum: 0, maximum: 100 }
        careerGapRisk: { type: integer, minimum: 0, maximum: 100 }
        upskillingNeeds: { type: integer, minimum: 0, maximum: 100 }
        overall: { type: integer, minimum: 0, maximum: 100 }
      required:
        - cvQuality
        - technicalSkillMatch
        - technicalRelevance
        - englishReadiness
        - europeJobReadiness
        - marketReadiness
        - careerGapRisk
        - upskillingNeeds
        - overall
    CandidateClassification:
      type: string
      enum: [recruiter_ready, needs_upskilling, needs_reskilling, not_ready_yet]
    SkillGap:
      type: object
      properties:
        targetRole: { type: string }
        required:
          type: array
          items: { type: string }
        matched:
          type: array
          items: { type: string }
        missing:
          type: array
          items: { type: string }
        matchPct: { type: integer, minimum: 0, maximum: 100 }
      required: [targetRole, required, matched, missing, matchPct]
    EvaluationInsight:
      type: object
      properties:
        title: { type: string }
        detail: { type: string }
        severity: { type: string, enum: [strength, opportunity, gap] }
      required: [title, detail, severity]
    Evaluation:
      type: object
      properties:
        candidateId: { type: string }
        evaluatedAt: { type: string, format: date-time }
        scores: { $ref: "#/components/schemas/EvaluationScores" }
        strengths:
          type: array
          items: { type: string }
        gaps:
          type: array
          items: { type: string }
        recommendedUpskilling:
          type: array
          items: { type: string }
        insights:
          type: array
          items: { $ref: "#/components/schemas/EvaluationInsight" }
        readinessTier: { type: string, enum: [emerging, developing, ready, elite] }
        classification: { $ref: "#/components/schemas/CandidateClassification" }
        skillGap: { $ref: "#/components/schemas/SkillGap" }
      required:
        - candidateId
        - evaluatedAt
        - scores
        - strengths
        - gaps
        - recommendedUpskilling
        - insights
        - readinessTier
        - classification
        - skillGap
    Candidate:
      type: object
      properties:
        id: { type: string }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        country: { type: string }
        targetRole: { type: string }
        yearsExperience: { type: integer }
        visaStatus: { $ref: "#/components/schemas/VisaStatus" }
        englishLevel: { $ref: "#/components/schemas/EnglishLevel" }
        euWorkEligible: { type: boolean }
        linkedinUrl: { type: string }
        createdAt: { type: string, format: date-time }
        cv:
          oneOf:
            - $ref: "#/components/schemas/CvUpload"
            - type: "null"
        evaluation:
          oneOf:
            - $ref: "#/components/schemas/Evaluation"
            - type: "null"
        skills:
          type: array
          items: { type: string }
        avatarUrl: { type: string }
        source: { type: string, enum: [direct, recruiter] }
        lastRole:
          oneOf:
            - { type: string }
            - { type: "null" }
        domain:
          oneOf:
            - { type: string }
            - { type: "null" }
        careerGapMonths: { type: integer, minimum: 0 }
        isShortlisted: { type: boolean }
        isClientReady: { type: boolean }
        isIndustryReady: { type: boolean }
        hasCvFile: { type: boolean }
      required:
        - id
        - fullName
        - email
        - phone
        - country
        - targetRole
        - yearsExperience
        - visaStatus
        - englishLevel
        - euWorkEligible
        - linkedinUrl
        - createdAt
        - skills
        - avatarUrl
        - source
        - careerGapMonths
        - isShortlisted
        - isClientReady
        - isIndustryReady
        - hasCvFile
...
    UserRole:
      type: string
      enum: [candidate, recruiter, admin]
    AuthUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        fullName: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        candidateId:
          oneOf:
            - { type: string }
            - { type: "null" }
      required: [id, email, fullName, role, candidateId]
    AuthSession:
      type: object
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
      required: [user]
    AuthRegisterRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        fullName: { type: string, minLength: 2 }
        role: { $ref: "#/components/schemas/UserRole" }
        gdprConsent: { type: boolean }
        candidateProfile:
          oneOf:
            - $ref: "#/components/schemas/CandidateRegistration"
            - type: "null"
      required: [email, password, fullName, role, gdprConsent]
    AuthLoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
...