# lib/api-zod/src/generated/types/candidateClassification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ORN-AI Talent Infrastructure Platform API
 * OpenAPI spec version: 0.1.0
 */

export type CandidateClassification =
  (typeof CandidateClassification)[keyof typeof CandidateClassification];

export const CandidateClassification = {
  recruiter_ready: "recruiter_ready",
  needs_upskilling: "needs_upskilling",
  needs_reskilling: "needs_reskilling",
  not_ready_yet: "not_ready_yet",
} as const;